Add extra configs and remove timeout default
This commit adds an `extra_configs` parameter to the channel
constructor. This parameter allows the user to be able to pass through
custom configuration parameters during channel creation, when necessary.

Previously, a 'timeout' value (the channel session timeout) could be
specified during channel creation, with a default value of 300 seconds.
At the streaming service, however, channel creation could fail if the
session timeout were greater than that of the request timeout.

This commit adds the ability to set the request and session timeouts
independently. The session timeout, if not set, now defaults to whatever
the streaming service would use (which should be less than that of the
default for the request timeout).