Reset stop state during stop before invoking stop callbacks
Previously, some of the channel stop state was only reset after
registered stop callbacks were invoked. If, during the execution of a
stop callback, a run() call was made, the run() call could lead to the
channel inadvertently being stopped again - with the original associated
stop callbacks fired recursively until the call stack was exhausted.

For the changes in this commit, the stop state is reset before any of
the stop callbacks are invoked. This allows for a run() call made during
the stop callback to correctly interpret that a second stop has not been
requested - avoiding the infinite recursive loop.